requirements.txt
langchain==0.0.267
sentence-transformers==2.2.2
qdrant-client>=1.11.0
pyyaml==6.0
transformers==4.31.0
torch==2.0.1
nltk==3.8.1
pydantic==1.10.9
pytest==7.4.0
colorlog==6.8.2
openai==1.30.1
tiktoken==0.7.0
langchain-community==0.2.12
langchain-opena==0.1.22